test(phpstan): enable PHPStan level 3 for `Pages/Ajax/`
Resolve PHPStan level 3 issues found when enabling checking of the
`Pages/Ajax/` directory.

+++ Pages/Ajax/ResourceDetailsPage.php
@@ -61,7 +61,7 @@ class ResourceDetailsPage extends Page implements IResourceDetailsPage
 
     public function GetResourceId()
     {
-        return ServiceLocator::GetServer()->GetQuerystring(QueryStringKeys::RESOURCE_ID);
+        return (int)ServiceLocator::GetServer()->GetQuerystring(QueryStringKeys::RESOURCE_ID);
     }
 
     /**
@@ -102,7 +102,7 @@ interface IResourceDetailsPage
 class ResourceDetailsPresenter
 {
     /**
-     * @var ResourceRepository
+     * @var IResourceRepository
      */
     private $resourceRepository;
